// src/types/productItem.ts
interface ProductItem {
    id: number;
    type: string;
    imageUrl: string;
    name: string;
    spec1: string;
    spec2: string;
    spec3: string;
    spec4: string;
    spec5: string;
    discount: string;
    specialOffer: string;
    originalPrice: string;
    affirmPrice: string;
    affirmUrl: string;
    shipping: string;
    arrivalTime: string;
    buyBtn: string;
}

export default ProductItem;